requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via starlette
black==24.10.0
    # via -r requirements/dev.in
cfgv==3.4.0
    # via pre-commit
click==8.1.8
    # via
    #   black
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime-gpu
dill==0.3.9
    # via multiprocess
distlib==0.3.9
    # via virtualenv
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
fastapi==0.115.6
    # via -r /home/geri/work/tools/ndi_recording/requirements/prod.in
filelock==3.16.1
    # via virtualenv
flake8==7.1.1
    # via -r requirements/dev.in
flatbuffers==24.12.23
    # via onnxruntime-gpu
h11==0.14.0
    # via uvicorn
humanfriendly==10.0
    # via coloredlogs
identify==2.6.6
    # via pre-commit
idna==3.10
    # via anyio
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via -r requirements/dev.in
mccabe==0.7.0
    # via flake8
mpmath==1.3.0
    # via sympy
multiprocess==0.70.17
    # via -r /home/geri/work/tools/ndi_recording/requirements/prod.in
mypy-extensions==1.0.0
    # via black
ndi-python==5.1.1.2
    # via -r /home/geri/work/tools/ndi_recording/requirements/prod.in
nodeenv==1.9.1
    # via pre-commit
numpy==2.2.2
    # via
    #   -r /home/geri/work/tools/ndi_recording/requirements/prod.in
    #   ndi-python
    #   onnxruntime-gpu
    #   opencv-python-headless
onnxruntime-gpu==1.20.1
    # via -r /home/geri/work/tools/ndi_recording/requirements/prod.in
opencv-python-headless==4.11.0.86
    # via -r /home/geri/work/tools/ndi_recording/requirements/prod.in
packaging==24.2
    # via
    #   black
    #   onnxruntime-gpu
    #   pytest
pathspec==0.12.1
    # via black
pillow==11.1.0
    # via -r /home/geri/work/tools/ndi_recording/requirements/prod.in
platformdirs==4.3.6
    # via
    #   black
    #   virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.1.0
    # via -r requirements/dev.in
protobuf==5.29.3
    # via onnxruntime-gpu
pycodestyle==2.12.1
    # via flake8
pydantic==2.10.5
    # via fastapi
pydantic-core==2.27.2
    # via pydantic
pyflakes==3.2.0
    # via flake8
pytest==8.3.4
    # via -r requirements/dev.in
pyyaml==6.0.2
    # via pre-commit
sniffio==1.3.1
    # via anyio
starlette==0.41.3
    # via fastapi
sympy==1.13.3
    # via onnxruntime-gpu
tomli==2.2.1
    # via
    #   black
    #   pytest
tqdm==4.67.1
    # via -r /home/geri/work/tools/ndi_recording/requirements/prod.in
typing-extensions==4.12.2
    # via
    #   anyio
    #   black
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   uvicorn
uvicorn==0.34.0
    # via -r /home/geri/work/tools/ndi_recording/requirements/prod.in
virtualenv==20.29.1
    # via pre-commit
